+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - `$BAT_CONFIG_DIR` is now a recognized environment variable. It has precedence over `$XDG_CONFIG_HOME`, see #1727 (@billrisher)
 - Support for `x:+delta` syntax in line ranges (e.g. `20:+10`). See  #1810 (@bojan88)
+- Add increment option in `--highlight-line` and `--line-range`, see #1611 (@t-mangoe)
 
 ## Bugfixes